Implement row actions feature
